requirements.txt
APScheduler==2.1.2
Cheetah==2.4.4
Flask==0.10.1
GnuPGInterface==0.3.2
Jinja2==2.7.2
Landscape-Client==13.07.3
M2Crypto==0.21.1
MarkupSafe==0.19
PAM==0.4.2
PyYAML==3.10
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
Werkzeug==0.9.4
apt-xapian-index==0.44
argparse==1.2.1
boto==2.2.2
chardet==2.0.1
cloud-init==0.6.3
command-not-found==0.2.44
configobj==4.7.2
distribute==0.7.3
euca2ools==2.0.0
gevent==1.0
greenlet==0.4.2
gunicorn==18.0
httplib2==0.7.2
itsdangerous==0.23
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
oauth==1.0.1
oauthlib==0.6.1
paramiko==1.7.7.1
pyOpenSSL==0.12
pycrypto==2.4.1
pycurl==7.19.0
pyserial==2.5
python-apt==0.8.3ubuntu7.2
python-debian==0.1.21ubuntu1
python-twitter==1.3.1
redis==2.9.1
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.3
ufw==0.31.1-1
unattended-upgrades==0.1
wadllib==1.3.0
wsgiref==0.1.2
zope.interface==3.6.1
